Support for large (2+ GB) recovery partitions
For use of NOOBS in other projects that use large images.

+++ recovery/imagewritethread.h
@@ -49,7 +49,7 @@ protected:
 signals:
     void error(const QString &msg);
     void statusUpdate(const QString &msg);
-    void parsedImagesize(unsigned int size);
+    void parsedImagesize(qint64 size);
     void completed();
 
 public slots:
